Removed un-needed variable 'firstPass' and moved getAssociatedPullRequestNumberFromSha outside of the GithubService class
