+++ lib/github-service.js
@@ -3,6 +3,14 @@ const { buildMarkdownText, buildPullRequestComment } = require('./utils');
 
 const hasCollectionSucceed = (assertions) => assertions.failed_count === 0;
 
+const getAssociatedPullRequestNumberFromSha = async (octokit,repository, commitSha) => {
+  return await octokit.request('GET /repos/{owner}/{repo}/commits/{commit_sha}/pulls', {
+    owner: repository.split('/')[0],
+    repo: repository.split('/')[1],
+    commit_sha: commitSha,
+  });
+}
+
 /**
  * Service responsible of calling Github and create the check containing postman collection results.
  */
@@ -29,18 +37,9 @@ class GithubService {
     });
   }
 
-  async getAssociatedPullRequestNumberFromSha(){
-    return await this.getOctokitInstance().request('GET /repos/{owner}/{repo}/commits/{commit_sha}/pulls', {
-      owner: this.options.repo.split('/')[0],
-      repo: this.options.repo.split('/')[1],
-      commit_sha: this.options.refCommit,
-    });
-  }
-
   async createPullRequestComment(report){
-    const res = await this.getAssociatedPullRequestNumberFromSha()
-    
-    if (res == null){
+    const res = await getAssociatedPullRequestNumberFromSha(this.getOctokitInstance(), this.options.repo, this.options.refCommit);
+    if (res.data.length == 0){
       console.log("[+] No pull request associated with commit sha")
       return; 
     }
