+++ lib/utils.js
@@ -73,7 +73,6 @@ const buildPullRequestComment = (report) => {
   .map((request, index) => {
     if (index == 0) {
       comment +=  " \n **The following tests are failing** \n";
-      firstPass = false;
     }
     comment += ( "\n [+] " + request.request_name + "\n");
     request.failed.map((test) => comment += "  -  " + test + " \n");
